requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.1
daphne==4.0.0
Django==4.1.7
frozenlist==1.3.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
multidict==6.0.4
numpy==1.24.2
opencv-python-headless==4.7.0.72
pika==1.3.1
Pillow==9.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.1.1
requests==2.28.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
yarl==1.8.2
zope.interface==6.0
